+++ python/cdp/openapi_client/api/payments_alpha_api.py
@@ -309,7 +309,6 @@ class PaymentsAlphaApi:
 
         # authentication setting
         _auth_settings: List[str] = [
-            'apiKeyAuth'
         ]
 
         return self.api_client.param_serialize(
@@ -591,7 +590,6 @@ class PaymentsAlphaApi:
 
         # authentication setting
         _auth_settings: List[str] = [
-            'apiKeyAuth'
         ]
 
         return self.api_client.param_serialize(
@@ -866,7 +864,6 @@ class PaymentsAlphaApi:
 
         # authentication setting
         _auth_settings: List[str] = [
-            'apiKeyAuth'
         ]
 
         return self.api_client.param_serialize(
@@ -1124,7 +1121,6 @@ class PaymentsAlphaApi:
 
         # authentication setting
         _auth_settings: List[str] = [
-            'apiKeyAuth'
         ]
 
         return self.api_client.param_serialize(
@@ -1403,7 +1399,6 @@ class PaymentsAlphaApi:
 
         # authentication setting
         _auth_settings: List[str] = [
-            'apiKeyAuth'
         ]
 
         return self.api_client.param_serialize(