+++ python/cdp/openapi_client/api/policy_engine_api.py
@@ -329,7 +329,6 @@ class PolicyEngineApi:
 
         # authentication setting
         _auth_settings: List[str] = [
-            'apiKeyAuth'
         ]
 
         return self.api_client.param_serialize(
@@ -626,7 +625,6 @@ class PolicyEngineApi:
 
         # authentication setting
         _auth_settings: List[str] = [
-            'apiKeyAuth'
         ]
 
         return self.api_client.param_serialize(
@@ -899,7 +897,6 @@ class PolicyEngineApi:
 
         # authentication setting
         _auth_settings: List[str] = [
-            'apiKeyAuth'
         ]
 
         return self.api_client.param_serialize(
@@ -1205,7 +1202,6 @@ class PolicyEngineApi:
 
         # authentication setting
         _auth_settings: List[str] = [
-            'apiKeyAuth'
         ]
 
         return self.api_client.param_serialize(
@@ -1530,7 +1526,6 @@ class PolicyEngineApi:
 
         # authentication setting
         _auth_settings: List[str] = [
-            'apiKeyAuth'
         ]
 
         return self.api_client.param_serialize(