+++ python/cdp/openapi_client/configuration.py
@@ -114,7 +114,6 @@ AuthSettings = TypedDict(
     "AuthSettings",
     {
         "apiKeyAuth": BearerFormatAuthSetting,
-        "endUserAuth": BearerFormatAuthSetting,
     },
     total=False,
 )
@@ -485,14 +484,6 @@ class Configuration:
         auth: AuthSettings = {}
         if self.access_token is not None:
             auth['apiKeyAuth'] = {
-                'type': 'bearer',
-                'in': 'header',
-                'format': 'JWT',
-                'key': 'Authorization',
-                'value': 'Bearer ' + self.access_token
-            }
-        if self.access_token is not None:
-            auth['endUserAuth'] = {
                 'type': 'bearer',
                 'in': 'header',
                 'format': 'JWT',