+++ python/cdp/openapi_client/models/sign_evm_transaction_criteria_inner.py
@@ -20,11 +20,12 @@ from pydantic import BaseModel, ConfigDict, Field, StrictStr, ValidationError, f
 from typing import Any, List, Optional
 from cdp.openapi_client.models.eth_value_criterion import EthValueCriterion
 from cdp.openapi_client.models.evm_address_criterion import EvmAddressCriterion
+from cdp.openapi_client.models.evm_data_criterion import EvmDataCriterion
 from pydantic import StrictStr, Field
 from typing import Union, List, Set, Optional, Dict
 from typing_extensions import Literal, Self
 
-SIGNEVMTRANSACTIONCRITERIAINNER_ONE_OF_SCHEMAS = ["EthValueCriterion", "EvmAddressCriterion"]
+SIGNEVMTRANSACTIONCRITERIAINNER_ONE_OF_SCHEMAS = ["EthValueCriterion", "EvmAddressCriterion", "EvmDataCriterion"]
 
 class SignEvmTransactionCriteriaInner(BaseModel):
     """
@@ -34,8 +35,10 @@ class SignEvmTransactionCriteriaInner(BaseModel):
     oneof_schema_1_validator: Optional[EthValueCriterion] = None
     # data type: EvmAddressCriterion
     oneof_schema_2_validator: Optional[EvmAddressCriterion] = None
-    actual_instance: Optional[Union[EthValueCriterion, EvmAddressCriterion]] = None
-    one_of_schemas: Set[str] = { "EthValueCriterion", "EvmAddressCriterion" }
+    # data type: EvmDataCriterion
+    oneof_schema_3_validator: Optional[EvmDataCriterion] = None
+    actual_instance: Optional[Union[EthValueCriterion, EvmAddressCriterion, EvmDataCriterion]] = None
+    one_of_schemas: Set[str] = { "EthValueCriterion", "EvmAddressCriterion", "EvmDataCriterion" }
 
     model_config = ConfigDict(
         validate_assignment=True,
@@ -68,12 +71,17 @@ class SignEvmTransactionCriteriaInner(BaseModel):
             error_messages.append(f"Error! Input type `{type(v)}` is not `EvmAddressCriterion`")
         else:
             match += 1
+        # validate data type: EvmDataCriterion
+        if not isinstance(v, EvmDataCriterion):
+            error_messages.append(f"Error! Input type `{type(v)}` is not `EvmDataCriterion`")
+        else:
+            match += 1
         if match > 1:
             # more than 1 match
-            raise ValueError("Multiple matches found when setting `actual_instance` in SignEvmTransactionCriteriaInner with oneOf schemas: EthValueCriterion, EvmAddressCriterion. Details: " + ", ".join(error_messages))
+            raise ValueError("Multiple matches found when setting `actual_instance` in SignEvmTransactionCriteriaInner with oneOf schemas: EthValueCriterion, EvmAddressCriterion, EvmDataCriterion. Details: " + ", ".join(error_messages))
         elif match == 0:
             # no match
-            raise ValueError("No match found when setting `actual_instance` in SignEvmTransactionCriteriaInner with oneOf schemas: EthValueCriterion, EvmAddressCriterion. Details: " + ", ".join(error_messages))
+            raise ValueError("No match found when setting `actual_instance` in SignEvmTransactionCriteriaInner with oneOf schemas: EthValueCriterion, EvmAddressCriterion, EvmDataCriterion. Details: " + ", ".join(error_messages))
         else:
             return v
 
@@ -100,13 +108,19 @@ class SignEvmTransactionCriteriaInner(BaseModel):
             match += 1
         except (ValidationError, ValueError) as e:
             error_messages.append(str(e))
+        # deserialize data into EvmDataCriterion
+        try:
+            instance.actual_instance = EvmDataCriterion.from_json(json_str)
+            match += 1
+        except (ValidationError, ValueError) as e:
+            error_messages.append(str(e))
 
         if match > 1:
             # more than 1 match
-            raise ValueError("Multiple matches found when deserializing the JSON string into SignEvmTransactionCriteriaInner with oneOf schemas: EthValueCriterion, EvmAddressCriterion. Details: " + ", ".join(error_messages))
+            raise ValueError("Multiple matches found when deserializing the JSON string into SignEvmTransactionCriteriaInner with oneOf schemas: EthValueCriterion, EvmAddressCriterion, EvmDataCriterion. Details: " + ", ".join(error_messages))
         elif match == 0:
             # no match
-            raise ValueError("No match found when deserializing the JSON string into SignEvmTransactionCriteriaInner with oneOf schemas: EthValueCriterion, EvmAddressCriterion. Details: " + ", ".join(error_messages))
+            raise ValueError("No match found when deserializing the JSON string into SignEvmTransactionCriteriaInner with oneOf schemas: EthValueCriterion, EvmAddressCriterion, EvmDataCriterion. Details: " + ", ".join(error_messages))
         else:
             return instance
 
@@ -120,7 +134,7 @@ class SignEvmTransactionCriteriaInner(BaseModel):
         else:
             return json.dumps(self.actual_instance)
 
-    def to_dict(self) -> Optional[Union[Dict[str, Any], EthValueCriterion, EvmAddressCriterion]]:
+    def to_dict(self) -> Optional[Union[Dict[str, Any], EthValueCriterion, EvmAddressCriterion, EvmDataCriterion]]:
         """Returns the dict representation of the actual instance"""
         if self.actual_instance is None:
             return None