+++ python/cdp/openapi_client/test/test_evm_smart_account.py
@@ -38,7 +38,9 @@ class TestEvmSmartAccount(unittest.TestCase):
             return EvmSmartAccount(
                 address = '0x742d35Cc6634C0532925a3b844Bc454e4438f44e',
                 owners = ["0xfc807D1bE4997e5C7B33E4d8D57e60c5b0f02B1a"],
-                name = 'my-smart-account'
+                name = 'my-smart-account',
+                created_at = '2025-03-25T12:00Z',
+                updated_at = '2025-03-26T12:00Z'
             )
         else:
             return EvmSmartAccount(