+++ python/cdp/openapi_client/test/test_import_evm_account_request.py
@@ -37,7 +37,8 @@ class TestImportEvmAccountRequest(unittest.TestCase):
         if include_optional:
             return ImportEvmAccountRequest(
                 encrypted_private_key = 'U2FsdGVkX1+vupppZksvRf5X5YgHq4+da+Q4qf51+Q4=',
-                name = 'my-wallet'
+                name = 'my-wallet',
+                account_policy = '123e4567-e89b-12d3-a456-426614174000'
             )
         else:
             return ImportEvmAccountRequest(