+++ typescript/src/CHANGELOG.md
@@ -1,5 +1,17 @@
 # CDP SDK Changelog
 
+## 1.18.0
+
+### Minor Changes
+
+- [#224](https://github.com/coinbase/cdp-sdk/pull/224) [`5d6e5a1`](https://github.com/coinbase/cdp-sdk/commit/5d6e5a164a23424444998cf51970836ee57ca659) Thanks [@0xRAG](https://github.com/0xRAG)! - Added useNetwork function on EvmServerAccount to enable node management and network hoisting
+
+## 1.17.0
+
+### Minor Changes
+
+- [#220](https://github.com/coinbase/cdp-sdk/pull/220) [`c91b8da`](https://github.com/coinbase/cdp-sdk/commit/c91b8da383f747de683019a011cbe3453684d6aa) Thanks [@sammccord](https://github.com/sammccord)! - Added support for `EvmDataCriterion` in policies, which can restrict smart contract interactions
+
 ## 1.16.0
 
 ### Minor Changes